backend/requirements.txt
celery==5.3.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Limiter==3.3.1
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
psycopg2_binary==2.9.7
redis==5.0.0
watchfiles==0.19.0
